Struct iapyx::controller::builder::ControllerBuilder

pub struct ControllerBuilder {
    backend: Option<ValgrindClient>,
    wallet: Option<Wallet>,
    settings: Option<Settings>,
}

Builder for wallet controller

Fields
backend: Option<ValgrindClient>wallet: Option<Wallet>settings: Option<Settings>
Implementations

impl ControllerBuilder

pub fn with_backend_from_address<S: Into<String>>( self, proxy_address: S, backend_settings: ValgrindSettings ) -> Result<Self, Error>

Sets backend from address along with connectivity setup (like cors/https)

Errors

On connectivity issues with given address

pub fn with_backend_from_client( self, backend: ValgrindClient ) -> Result<Self, Error>

Sets backend from ValgrindClient struct

Errors

On connectivity issues with given address

pub fn with_wallet_from_secret_file<P: AsRef<Path>>( self, private_key: P ) -> Result<Self, Error>

Sets wallet based on secret file with bech32 secret key

Errors

On incorrect format of private key (it should be bech32), problem with reading file or parsing wallet crypto material

pub fn with_wallet_from_secret_key( self, private_key: SecretKey<Ed25519Extended> ) -> Result<Self, Error>

Sets wallet based on secret key struct

Errors

On incorrect format of private key (it should be bech32) or parsing wallet crypto material

pub fn with_wallet_from_qr_file<P: AsRef<Path>>( self, qr: P, password: &str ) -> Result<Self, Error>

Sets wallet based on qr code with private key

Errors

On incorrect qr code, problem with reading file or parsing wallet crypto material

Panics

In unlikely event of 10 not fit in u8 type

pub fn with_wallet_from_account(self, account: &[u8]) -> Result<Self, Error>

Sets wallet based on account bytes

Errors

On parsing wallet crypto material

pub fn with_wallet_from_qr_hash_file<P: AsRef<Path>>( self, qr_payload_file: P, password: &str ) -> Result<Self, Error>

Sets wallet based on payload file (in form of hex encoded string) with private key material

Errors

On reading file or parsing wallet crypto material

pub fn build(self) -> Result<Controller, Error>

Builds Controller

Errors

On any missing configuration steps