Struct iapyx::controller::Controller

pub struct Controller {
    pub backend: ValgrindClient,
    pub wallet: Wallet,
    pub settings: Settings,
    pub block_date_generator: BlockDateGenerator,
}

Responsible for all wallet operations from voting to retrieving information about proposals or voting power

Fields
backend: ValgrindClient

Catalyst backend client

wallet: Wallet

Wallet state

settings: Settings

Cached blockchain settings

block_date_generator: BlockDateGenerator

Expiry date generator

Implementations

impl Controller

pub fn switch_backend( &mut self, proxy_address: String, backend_settings: RestSettings ) -> Result<(), ControllerError>

Connects to new backend

Errors

On connectivity issues

pub fn account(&self, discrimination: Discrimination) -> Address

Gets account

pub fn id(&self) -> AccountId

Gets account identifier. This is something different that account method since we are not retrieving entire object but only the identifier which helps of find records in blockchain

pub fn send_fragment( &self, transaction: &[u8] ) -> Result<FragmentId, ControllerError>

Sends raw transaction bytes

Errors

On connection issues

Panics

On internal error in which vec does not hold transaction

pub fn set_block_date_generator( &mut self, block_date_generator: BlockDateGenerator )

Sets transaction ttl definition. Usually when we are sending some fragments we need to define how long we want to wait until it will be put in block. BlockDateGenerator helps us define ttl without calculating it each time

pub fn send_fragments( &self, transaction: Vec<Vec<u8>> ) -> Result<Vec<FragmentId>, ControllerError>

Sends collection of raw transaction bytes

Errors

On connection issues

pub fn confirm_all_transactions(&mut self)

Confirms all transactions means to remove them from pending transaction collection and as a result remove trace needed to track their status in node

pub fn confirm_transaction(&mut self, id: FragmentId)

Confirms transaction by it id. This means to remove it from pending transaction collection and as a result remove trace needed to track their status in node

pub fn pending_transactions(&self) -> Vec<FragmentId>

Unconfirmed collection of transactions (which statuses we still want to track)

pub fn wait_for_pending_transactions( &mut self, pace: Duration ) -> Result<(), ControllerError>

Waits until all transaction will have final states (either InABlock or Reject) for given duration

Errors

On connection issues

pub fn remove_pending_transaction(&mut self, id: FragmentId)

remove specific transaction from assumed pending

pub fn total_value(&self) -> Value

gets total value ada

pub fn refresh_state(&mut self) -> Result<(), ControllerError>

Reload wallet state from blockchain

Errors

On connection issues

pub fn get_account_state(&self) -> Result<AccountState, ControllerError>

Get account state from blockchain (ada/delegation status etc.)

Errors

On connection issues

pub fn proposals( &self, group: &str ) -> Result<Vec<FullProposalInfo>, ControllerError>

Gets proposals from vit-servicing-station

Errors

On connection issues

pub fn funds(&self) -> Result<Fund, ControllerError>

Gets proposals from vit-servicing-station

Errors

On connection issues

pub fn settings(&self) -> Result<SettingsDto, ControllerError>

Gets blockchain Settings from blockchain

Errors

On connection issues

pub fn vote_for( &mut self, voteplan_id: &str, proposal_index: u32, choice: u8 ) -> Result<FragmentId, ControllerError>

Send specialized transaction (with vote certificates) based on low level parameters

Errors

On connection issues

pub fn vote( &mut self, proposal: &FullProposalInfo, choice: Choice ) -> Result<FragmentId, ControllerError>

Send specialized transaction (with vote certificates) based on FullProposalInfo struct

Errors

On connection issues

pub fn votes_batch( &mut self, votes_data: Vec<(&FullProposalInfo, Choice)> ) -> Result<Vec<FragmentId>, ControllerError>

Send collection for specialized transaction (with vote certificates) based on FullProposalInfo struct

Errors

On connection issues

Panics

On internal error when updating wallet

pub fn get_proposals( &mut self, group: &str ) -> Result<Vec<FullProposalInfo>, ControllerError>

Gets proposals from vit-servicing-station based on voting group

Errors

On connection issues

pub fn fragment_logs( &self ) -> Result<HashMap<FragmentId, FragmentLog>, ControllerError>

Gets fragment logs from node

Errors

On connection issues

pub fn vote_plan_history( &self, vote_plan_id: VotePlanId ) -> Result<Option<Vec<u8>>, ControllerError>

Gets vote plan history from node for given vote plan id

Errors

On connection issues

pub fn votes_history( &self ) -> Result<Option<Vec<AccountVotes>>, ControllerError>

Gets vote plan history from node

Errors

On connection issues

fn wallet_address(&self) -> Result<Address, ControllerError>

pub fn active_vote_plan(&self) -> Result<Vec<VotePlanStatus>, ControllerError>

Gets vote plan from node

Errors

On connection issues