Struct iapyx::MultiController

pub struct MultiController { /* private fields */ }

Responsible for controlling more than one wallet at one time. Useful for load scenario or wallets which handles many users

Implementations

impl MultiController

pub fn recover_from_qrs<P: AsRef<Path>>(
wallet_backend_address: &str,
qrs: &[P],
pin_mode: &PinReadModeSettings,
backend_settings: RestSettings
) -> Result<Self, MultiControllerError>

Creates object based on qr codes files

Errors

On backend connectivity issues or parsing qr problems

Panics

On internal error when exposing secret key

pub fn recover_from_sks<P: AsRef<Path>>(
proxy_address: &str,
private_keys: &[P],
backend_settings: RestSettings
) -> Result<Self, MultiControllerError>

Creates object based on secret files

Errors

On backend connectivity issues or parsing qr problems

Panics

On single wallet recover error

pub fn proposals(
&self,
group: &str
) -> Result<Vec<FullProposalInfo>, MultiControllerError>

Gets proposals from vit-servicing-station based on voting group

Errors

On connection issues

pub fn update_wallets_state(&mut self) -> Result<(), MultiControllerError>

Update wallet state for entire wallets collection based on current state in blockchain

Errors

On connection issues

pub fn update_wallet_state(
&mut self,
wallet_index: usize
) -> Result<(), MultiControllerError>

Update wallet states based on current state in blockchain

Errors

On connection issues

pub fn update_wallet_state_if(
&mut self,
wallet_index: usize,
predicate: &dyn Fn(&Wallet) -> bool
) -> Result<(), MultiControllerError>

Update wallet states based on current state in blockchain if condition is satisfied. For example only if wallet spending counter is 0 which means it is in initial state

Errors

On connection issues

pub fn vote(
&mut self,
wallet_index: usize,
proposal: &FullProposalInfo,
choice: Choice,
valid_until: BlockDate
) -> Result<FragmentId, MultiControllerError>

Sends vote transaction on behalf of wallet with index wallet_index on proposal with given choice. Sets expiry slot equal to valid_until.

Errors

On connection issues

pub fn votes_batch(
&mut self,
wallet_index: usize,
use_v1: bool,
votes_data: Vec<(&FullProposalInfo, Choice)>,
valid_until: &BlockDate
) -> Result<Vec<FragmentId>, MultiControllerError>

Sends bunch of vote transactions on behalf of wallet with index wallet_index with map of proposals and respectful choices. Sets expiry slot equal to valid_until. Method can use V0 or V1 api based on preference. V1 enable to send all transactions in a batch as single call whether legacy V0 is able to send them as one vote per call

Errors

On connection issues

Panics

On connection problem

pub fn confirm_all_transactions(&mut self)

Confirms all transactions for all wallets. Confirms means loose interest in tracking their statuses. This method should be call when we are sure our transactions all in final states (in block or failed).

pub fn confirm_transaction(&mut self, fragment_id: FragmentId)

Confirms all transactions for wallet. Confirms means loose interest in tracking their statuses. This method should be call when we are sure our transactions all in final states (in block or failed).

pub fn wallet_count(&self) -> usize

Wallet counts